requirements.txt
pip
motor
pymongo[srv]
praw
lxml
python-codeforces
discord.py
pytz
periodic
wikipedia
googletrans
udpy
pillow
discord.py[voice]
yt-dlp
youtube-dl
DiscordUtils
youtube-title-parse
molmass
mendeleev
SpotifyUtil
pytube
num2words
grequests
python-aternos
js2py